Invalidate edited commits messages for edited stack on confirm
Summary:
Finally, we can fix the bug this stack has been about.

When we go to confirm the split/edit stack, we now know the user has confirmed or cleared their unsaved message changes via the modal. We've also now loaded the edited commit message into the edit stack data structure. This means `debugimportstack` has now imported those edited messages.

Now all that's left is that we need to prevent the succession of the stack from resurrecting the edits once again. Since they've been imported now, it's safe to discard all the unsaved edits for the entire stack.

You may think that this is unnecessary because we already have the same message, so it wouldn't detect the message as being different. This is only true for 1:1 predecessor:successor changes and falls apart with non-1:1 edits like split and fold.

The only tiny edge case is that if the split fails, we've now cleared their message without it getting to get saved. I think this should be an unlikely case—debugimportstack is designed not to fail as much as possible (we should instead fail earlier at the debugexportstack stage, which is safe)

Reviewed By: quark-zju

Differential Revision: D49551311

fbshipit-source-id: dadccda9b462a47d9b4bbdd48cdeb519f5df5672