Add counters to Overlay to track hits and misses
Summary:
We would like to know how effective our various caching entities are. Here we are adding counters in Overlay to track hits and misses when loading an overlay directory.

File operations in the Overaly are only used when the overlay checker is being ran - as such not adding counters for those.

Reviewed By: MichaelCuevas

Differential Revision: D49552392

fbshipit-source-id: 1d2e699b4b9c5a1bad3d47ef1e7945a5f7f37999