dispatch: log structured cli arg info (Python)
Summary:
Log the same fields as Rust dispatch code. Now we will have structured info for all options/args.

The intention is to use a low retention for the positional args and option values since those will have high cardinality, but keep the option names around longer. The option names in particular are useful for checking where/when particular CLI flags are in use (as opposed to trying to regex match against the unstructured "fullcommand").

Reviewed By: zzl0

Differential Revision: D49509123

fbshipit-source-id: d95c1e2cea46534247b6c3a981d7a48a0d7d17c9