metrics: log structured cli arg info
Summary: Now we log positional args, option names, and option values to the "sampling" file. The intention is to use a low retention for the positional args and option values since those will have high cardinality, but keep the option names around longer. The option names in particular are useful for checking where/when particular CLI flags are in use (as opposed to trying to regex match against the unstructured "fullcommand").

Reviewed By: quark-zju

Differential Revision: D49509126

fbshipit-source-id: 4f5803a084de74d68df66356fd8eaf11fa3e9dff