mononoke_app/args: Support passing multiple changesets
Summary:
Later in this stack, we port `mononoke_admin derived-data` to `mononoke_newadmin`.

One functionality that doesn't currently exist in the `mononoke_app` framework yet, and that `derived-data` currently relies on is that of passing multiple changesets to a CLI tool.
We judge it would be a UX regression to limit this app to passing a single changeset at a time.

`derived-data` uses `resolve_csids`, which also gives mix-and-match functionality between different ways to describe changesets in the same list.
We deliberately don't support this because it would make the implementation less idiomatic, using old clap constructs to preserve order, and offer a flexibility that is not necessary from a user's perspective.

Reviewed By: markbt

Differential Revision: D49271578

fbshipit-source-id: fd72a885c36bd4c4fc810d821dcd75324d00fe81