Generate ClientRequestInfo if missing
Summary: For initiator services, if `ClientRequestInfo` is missing, we need to generate it. This diff adds support for that. Note that even if the service is not an initiator service (e.g. `LandService`), we populate `ClientRequestInfo` with `ClientEntryPoint` set as `LandService`. Ideally this scenario should never happen since the caller should always be a known entry point. Presence of such entries in the scuba logs will indicate fishy business that we need to investigate and fix.

Reviewed By: liubov-dmitrieva

Differential Revision: D49485744

fbshipit-source-id: f65aa7e9e6462dd7b92f122a6e7125b62d7eb0cf