Added default summary/test-plan templates for IGB engs
Summary:
as a part of this initiative https://fburl.com/workplace/ki1ksbyc IG Business is exploring possibility of using templates for summary / test plans to empower authors to write better diffs and make them easier to review.

this diff introduces custom default values for those fields and uses gatekeeper-based gating to roll it out for a limited audience.

Reviewed By: evangrayk

Differential Revision: D49484684

fbshipit-source-id: d8544e393d858b07e111ddb6de1b41709d92cdf8