status: use up-to-date dirstate p1 in eden request
Summary: During multi-step operations like rebase, we move the working copy to different parents. Previously the Rust workingcopy's "status" calls to EdenFS were not using an up-to-date parent id in the "status" args, resulting in the "requested parent commit is out-of-date" error. Fix by pulling the parent id out of the treestate for every "status" call.

Reviewed By: quark-zju

Differential Revision: D49509139

fbshipit-source-id: 7dff48e826c073897c5fd1f683ac76f2f0068753