Add integration test for renamed export path
Summary:
Adding an integration test case to repro the issue currently blocking gitexport.

The problem is that, even though we properly set the new parents in the bonsai changeset, there are still references to the old parents in the `copy_from` field of the `TrackedFileChange`.

Example: we export directory `foo` and have the following 3 changesets `A <- B <- C`, where
- A changes the directory `bar`
- B changes a random directory
- C renames `bar` to `foo`

This would crash because `C` will have file changes referencing `B`.
This will be fixed in the next few diffs by properly updating those references the same way we update the parents.

**Note:** I added two calls to `gitexport` on the test: one passing the old path name manually and another without doing that. That's because the former will work with some small tweaks that will unblock the use of the tool while the proper fix is under development.

Differential Revision: D49366697

fbshipit-source-id: c13b76abcee5f780c4a0f46e65e047ef4f8745d8