Get ChangesetPathHistoryContext from each path individually
Summary:
## Context
D49366697 adds an integration test to cover a bug that was surfaced when running the tool in a real directory.

The **TL;DR** of the bug is that gitexport crashes when we pass `foo` as the export directory and at some point in history `foo` was created by renaming an entire directory named `bar`.

## Temporary solution
Manually pass `bar` as an export directory.

However, this doesn't work out of the box because of the way that the `paths_with_history` method works.

The documentation of `paths_with_history` says that it will not basically ignore any paths that don't exist in the changeset from which the method was called.

This means that passing `bar` as an export directory has no effect, since it doesn't exist in the starting changeset anymore.

By calling `path_with_history` for each path, we'll get the changesets affecting the old directories, even if they don't exist in the starting changeset anymore.

This unblocks the fix in D49416858.

Differential Revision: D49369883

fbshipit-source-id: dc3a44752cdcf76f03efdcd5e0e76a1435972026