Set new parent in copy_from file changes
Summary:
## Bug
See D49366697 and D49369883 for context on the bug.

## Solution
The comments on the code should explain the solution (please LMK if that's not the case).
The **TL;DR** is that any files that are copied  in a changeset `A` should only be copied from a changeset that are is being exported to the git repo, and thus have been rewritten before `A` and will be present in the `remapped_parents` map.

## What's next
Stop relying on the temporary hack of manually passing the name of the old directory as an export path by automatically getting the name of the old directories and all the changesets that affected it.

Differential Revision: D49416858

fbshipit-source-id: d084427a2db0c835ddb6ba39c4148bd2a4df82f7