hooks: export error type precisely
Summary:
Use of `*` in imports is discouraged, and doubly so when the import is re-exported with `pub use`.  We only want to export the error type, so specify exactly that.

This fixes a new warning that shows up in the latest compiler version about the shadowing of `NonRootMPath` through the blanket import.

Differential Revision: D49680495

fbshipit-source-id: 7fd939d2a7c9e9a42e3f6d31679564a7aea7fcd0