Use purge instead of a manual delete file function
Summary:
Apparently `sl purge` is capable of deleting individual files. This is much nicer than what we did before, which was to just run `rm` ourselves. By running this through `sl`, we can queue it up, show optimistic state, handle cancelling, etc.

Plus, it's nice from a security perspective to not have an exposed way to delete arbitrary paths. I think we handled the security correctly, but it's better to not have to think about it at all.

Reviewed By: quark-zju

Differential Revision: D49616814

fbshipit-source-id: 8e1e4bef34310d357f4eaf7b2b38a8cf1932560e