pymodules: use ModuleSpec from importlib
Summary:
Previously, we defined our own `ModuleSpec` to avoid importing the `importlib`
on disk. That can cause issues with code that wants to set random attributes
on the `ModuleSpec` (heap) object. For example, Python 3.11 wants to set the
`_uninitialized_submodules` attribute, which fails with the current pymodules
implementation.

Actually, we can get the `ModuleSpec` from the frozen `_frozen_importlib`.
That does not hit the filesystem. Switch to the stdlib `ModuleSpec` for better
compatibility.

Reviewed By: sggutier

Differential Revision: D49710052

fbshipit-source-id: e0e5ad0267e29d659da25d167f9e7254adf198b9