warm_bookmarks_cache: always find the latest underived bookmark value
Summary:
`find_latest_derived_and_oldest_underived` is both misnamed, and doesn't work quite right.

First, it actually wants to return the latest underived commit, not the oldest one, as we will attempt to derive everything that needs to be derived in a single batch via the derived data service.

Second, when we step back farther than the original set of 10 entries, we don't want to re-set the value of `latest_underived.maybe_cs_id`, as we still want the first commit that we original found to ensure that we always make progress.  Note that we can't use the `None` value of this type as that means "deleted", so instead introduce a tracking bool.

Reviewed By: mitrandir77

Differential Revision: D49688507

fbshipit-source-id: 446d7677125b809f2aa27cca5c99ac9f854f4024