Reset selection when using goto button
Summary:
Small feature that we sort of had in the old ISL. After clicking "goto", we can reset your commit selection so that the new destination is implicitly shown in the sidebar.

We don't need to actively select your new destination, since we show the head commit by default and it's the new head commit already. This is a bit nicer because then we don't need the selection highlight to appear on your head commit, which is just visual noise unless you intentionally are clicking to select.

Reviewed By: zzl0

Differential Revision: D49613263

fbshipit-source-id: be763b8b48218038996620c6d5d37bf53df2f5b5