Use position sticky for file names in chunk selection
Summary:
This is a trick we do in the comparison view, where we make the filename at the top of each diff view sticky, so that you can always see the filename as you scroll. `position: sticky` works quite well "for free" here.

This way, if you expand a chunk selection, you can see the current filename as you scroll.

Note that we also need to  consider how this looks if you have lots of unexpanded files. It does look slightly different, since the first file will stick a bit. But I think it looks perfectly reasonable, and it makes the chunk selection use case much nicer.

Reviewed By: zzl0

Differential Revision: D49614343

fbshipit-source-id: 3f7e93f1c2d992c401fc4898b5f860b173ecacb6