fail startup if the localstore fails to open
Summary:
Currently EdenFS is swallowing errors opening the local store. None of the
mounts will be mounted, but EdenFS will report that start was successful.
Previously, errors opening the local store would cause startup to fail.
D38190050 accidentally caused this behavior.

We want to go back to EdenFS more explictly failing to start because this is
clearer to users, and EdenFS will not be able to work with out it's local store.

I discovered a potential cause of graceful restart leaving behind stale mounts
while putting the start failure behavior back:

If EdenFS fails in startup sometime between receiveing takeover data and
creating setting up a mount object, that mount will be left stale. We probably
don't want that, but fixing that is more complicated, and it's also very old
behavior (happened before D38190050), so I'm going to delay the fix on that.

Reviewed By: MichaelCuevas

Differential Revision: D48813074

fbshipit-source-id: 3c129c48b423afb2b0224f1eec1031c26c8968c6