symlinks: resolve issue with status and symlinks disabled
Summary: Some users reported seeing files that were supposed to be symlinks when having symlinks disabled. This was caused by the initial mode for those "symlinks" being symlinks; with symlinks disabled this symlink mode was supposed to be ignored and the file should have been treated as a regular file.

Reviewed By: kmancini

Differential Revision: D49610768

fbshipit-source-id: 73b78cc1b5f278ac2dd871bf9454fec63c723979