hook_manager: extract hook manager to its own crate
Summary: The hook manager is co-located with the hooks it manages, making it hard to separate them out and test them.  Split it off to its own crate.  The file content providers that don't depend on the repo itself can also be moved in with the manager, allowing the manager-specific tests to be located with the manager.  The tests in the `hooks-tests` crate that correspond to testing the manager specifically are moved into the manager crate.

Reviewed By: RajivTS

Differential Revision: D49094390

fbshipit-source-id: c4fa0f10c503c35aafb62b4cb54685f797e50fd4