hooks: extract repo_hook_file_content_provider to its own crate
Summary: `RepoHookFileContentProvider` is separate from other content providers, as it depends on repo attributes to provide the content, and we wish to avoid these dependencies within the hook manager.  Move this into its own crate alongside the hook manager, along with the tests that are specific to it from the `hooks-tests` crate.

Reviewed By: mitrandir77

Differential Revision: D49094389

fbshipit-source-id: 8cb7b9825216b00037a96a1e69fe76fa83d8d66b