Fix syntax highlighting not working in non-dev browser platforms
Summary:
Syntax highlighting was working in the browser if you ran in dev mode. It also worked in vscode in production.

However, apparently I wasn't thorough in testing this for production browser ISL instances.

We have an asset manifest which controls which files we server in our server. This is better for security. However, in our dev mode server, we don't use this file serving, and instead files are served (and hot reload from) CRA's dev server, which means this asset checking is not done.

In production, we don't use the CRA dev server, so the requests fail the asset manifest check.

We need to be able to load the generated files for syntax highlighting. It's not enough to merely be in the public/ folder.

To do this, we can simply insert these files into our asset manifest object when setting up the server.

This fix will also apply to other places where we use the server code, including Android Studio, browser, chromelike app, and webview.

Reviewed By: quark-zju

Differential Revision: D49756350

fbshipit-source-id: 9b70f65e6741593de83025229d958c0288289711