Modify BonsaiTagMapping schema to accomodate Tag ObjectId
Summary:
Currently, we record Git tags as bookmarks and while adding such bookmarks we keep track of the tag-name, the bonsai changeset that the tag points to and the bonsai changeset that contains the metadata about the tag. We also separately store the raw Git tag object but do not store any references to it in any mapping table. Hence, when we need to include these tags in packfiles or bundles, we cannot reach them. This diff modifies the `bonsai_tag_mapping` table to track the `ObjectId` corresponding to the Git tag object. Since the users of these table have not been modified (yet) to populate this new field, the `tag_hash` column has a default value of `x00000000000000000000`.

I will also run AOSC modify once this diff is landed to make sure the changes are reflected.

Differential Revision: D49804682

fbshipit-source-id: 4eb45cb08dbefec612db26db2373435143be3dca