Support delta variants as packfile item
Summary: Now that we have support for deltas, the `PackWriter` and `BundleWriter` should support them as well. This diff updates the types of `PackfileItem`s by adding `OidDelta` variant. The `RefDelta` variant is not included since we do not use it. The packer and bundler would now take a stream of `PackfileItem`s instead of a stream of `Bytes` and would require the caller to construct the appropriate type of `PackfileItem` from input bytes

Differential Revision: D49270575

fbshipit-source-id: 9bed04766151e15d1d31d666f2f93a85c269b6ec