Update DeltaInstructions storage logic
Summary:
`DeltaInstructions` can potentially be large in size even though they are chunked. Currently, the instructions are stored as-is and it is at fetch time that we need to Zlib encode these instructions so that Git understands them. Instead of paying the penalty fetch time (which will be used in multiple memory and latency sensitive services), we can encode the instructions while they are being written to the blobstore (in derivation worker). Since we never have any use for the decoded instructions, it is safe to have encoding at rest.

As part of this diff, I also added a method to fetch the chunked instructions as a stream of bytes which is currently only used in test but would later be used for production use-case. Additionally, I have included two other properties in `ObjectDelta` which represent the size of the delta instructions with and without encoding. The raw size is necessary metadata that we need to include in packfiles and bundles and the compressed_size is useful to determine if we shoud use the delta or use the raw object directly.

Differential Revision: D49227944

fbshipit-source-id: 470789a1b708751d0e26a4f278350f5513a51fd4