config: show invalid/dropped config values
Summary:
If config values aren't in configs.allowedconfigs and their containing file isn't in configs.allowedlocations, we drop the config values. Previously there was no way to see them using "sl config", but now they will show up with an "(invalid - dropped)" source.

We achieve this by not actually deleting the config values. If the values are before the active value, we leave them since they will have no effect. If they are after the config value, we scooch them before the active value so they have no effect. If there is no active value, we append a dummy "unset" value so we can keep the invalid values.

Reviewed By: quark-zju

Differential Revision: D49274810

fbshipit-source-id: c2b078718b7bdebec825dc4c382e2b7f8c962988