restack: fix SRC placeholder interaction w/ titles namespace
Summary:
Restack uses the "SRC" placeholder internally as a template variable to be expanded for each commit being rebased. This was colliding with the "titles" namespace if any draft commits contained "src" in their titles.

Do a quick fix to avoid fast-path scmutil.revsingle if the "dest" revset contains "SRC". Maybe a better fix is to keep user-specified and restack-specified revsets separate, but this fix seemed easier for now.

Reviewed By: quark-zju

Differential Revision: D49754265

fbshipit-source-id: f0440d85e0baa83b3639d848d094fc41a1b37922