set ClientRequestInfo in ObjectFetchContext requestInfo
Summary: If set and non-empty, sets the `ClientRequestInfo` fields in `ObjectFetchContext`'s requestInfo map. This is not consumed anywhere yet, but will passed to the `sapling_backing_store` in follow up diffs. This is part of a broader effort to thread a common client identifier from scm clients (`sapling` and `edenfs` to Mononoke)

Reviewed By: jdelliot

Differential Revision: D49716011

fbshipit-source-id: 1de0044f9c516c9deb4ef82ff269a3909d87a347