Update create_git_tag endpoint to allow for tag hash
Summary: `remote-gitimport` makes use of `scs` to record git tags at Mononoke end. This diff updates the parameter type for the `create_git_tag` endpoint to allow for specifying the hash of the git tag being created. It is optional since the callers will be updated later to provide the tag hash and meanwhile a none value will be used.

Differential Revision: D49805555

fbshipit-source-id: 25c295183cdfdd2259add84a79518a106f210e5b