Update BonsaiTagMapping facet to handle tag hash
Summary: In the previous diff, the schema of the `bonsai_tag_mapping` table was modified to accomodate the hash of the Git tag object. This diff updates the logic in the `bonsai_tag_mapping` crate to handle the hash while inserting and retrieving entries. It also changes the `add` behavior to `add_or_update` behavior to ensure that the object corresponding to the tag-name can be modified since at the Git end the user can delete and recreate the same name tag but with different properties.

Differential Revision: D49804791

fbshipit-source-id: 8b6776d538dabea9ada1ad26d7c9269010b35652