Add tag hash support in gitimport
Summary: This diff implements the necessary changes to ensure that when direct-`gitimport`ing a repo, the imported tags are recorded along with the tag hash of the Git object ID of the tag. Note that this just applies for direct `gitimport`. Support for `remote-gitimport` would be added in a follow up diff.

Reviewed By: mzr

Differential Revision: D49805885

fbshipit-source-id: 18502476298acaf9f04ab76c9de9f395018b9247