derivation_queue: add enqueue time and log time spent in queue
Summary:
Now that we can pass additional information from the derived data service to the derivation workers, start by passing the enqueue time and logging the approximate time spent on the queue.

Note that this is approximate as we rely on the clocks on the derived data service and derivation workers being in sync.

Reviewed By: mitrandir77

Differential Revision: D49689667

fbshipit-source-id: 86c09bc8c8800f10cb25c73831619bb7f5ee3bfe