derivation_worker: log client request info to scuba
Summary: Add logging of client request info objects to scuba in the derviation worker.  This requires us to stop using a single context object, and instead construct a new one for each request.  For now we just create a new client info.  We will accept one from the derived data service in a later diff.

Reviewed By: RajivTS

Differential Revision: D49743404

fbshipit-source-id: 68b16b265593a8a4f487976f96e4d3e62904b543