thread ObjectFetchContext to HgDatapackStore
Summary: In order to pass the requestInfo map to deeper layers, thread the Thrift layer `ObjectFetchContext` to `HgDatapackStore`. This is not consumed anywhere yet, but will passed to the sapling_backing_store in follow up diffs. This is part of a broader effort to thread a common client identifier from scm clients (sapling and edenfs to Mononoke)

Reviewed By: jdelliot

Differential Revision: D49795940

fbshipit-source-id: 67b2f703caae403f9d93fcb010b810a61437017d