update: remove experimental.updatecheck config
Summary:
Been deprecated for a while in favor of "commands.update.check".

There is a terrible subtlety here: "commands" config settings are automatically filtered out when HGPLAIN=1. "experimental" settings are not. So, experimental.updatecheck=noconflict was active with HGPLAIN=1, but commands.update.check=noconflict is not active with HGPLAIN=1. This would be a change in behavior, though, so I set the configitems.py default value to "noconflict" which will take effect with HGPLAIN=1.

Still, there is a change in behavior in that locally configured commands.update.check behavior will no longer take effect with HGPLAIN=1. It doesn't look to be configured anywhere outside hg, so I think we are okay.

We are planning to simplify and improve the update check with upcoming Rust work, so this config knob should go away, hopefully.

Reviewed By: sggutier

Differential Revision: D49058007

fbshipit-source-id: af2d01bf7bc6fac0bbbd913b745ce5da8353f13d