Implement support for RefDelta to OffsetDelta conversion during packing/bundling
Summary: As part of `PackfileItem`, we support only two variants, `Base` where the direct Git object is used and `OidDelta` where the delta is expressed w.r.t a base object specified through its Object Id. Depending upon the type of pack/bundle being generated, we might want to have deltas only in the form of offset (i.e. deltas where the base object is present at an offset from the current object in the same packfile). This diff adds support for converting `OidDelta` into `OffsetDelta` directly when writing to the pack so that the input can continue to be either `Base` or `OidDelta`. This is necessary since the input is generated through `GitDeltaManifest` and its not feasible to generate `OffsetDelta` at that stage without sequentializing the entire processing thereby slowing it down considerably.

Reviewed By: markbt

Differential Revision: D49863590

fbshipit-source-id: 8501d3b22c77bbba2c840bd96066ca6098a562bd