Add telemetry to HgImporter to record failures & errors
Summary:
As part of the project to remove HgImporter adding some basic telemetry for failures returned from HgImporters. This will pair up with the remaining errors returned from EdenApiMiss (which is more accurately described as backing store miss).

This should not overwhelm any collectors as it comes after we have removed the highest contributor of edenapi_miss events. That said, once landed we can contrain the data retention of this event (1 day/50GB).

Reviewed By: xavierd

Differential Revision: D49935347

fbshipit-source-id: 92d815d802d083181a9cbc74e9df025b611e1d4c