Fix issue with HgProxyHash::getBatch where results are out of order
Summary:
When `HgQueuedBackingStore::prefetchBlobs` invokes `HgProxyHash::getBatch` it expects the results to be returned in the same order at the request. See https://fburl.com/code/bhca8zn9. However, the code, as written today, can re-order the results if there are any hashes loaded from the LocalStore.

It is assumed there are still, in existence, some clones that have these older proxy hashes stored in LocalStore. For those clones, if the user were to prefetch blob data, the results would be incorrect.

This change ensure that the ordering of the results is the same as the requests. In the domminante use case - where all all OjbectsIds are found to be embedded proxy hashes there will be a small slowdown in transforming the results from variants to HgProxyHashes. This is not expected to be a significant cost. In the case where there are mixed results or only hashes stored in LocalStore, there is a similar overhead, but this should be even less impactful.

Reviewed By: xavierd

Differential Revision: D49923147

fbshipit-source-id: 489c74945cd2ad33e84cf62e92f8e005aed6a05c