a hacky way to separate inode and tree evaluate in GlobNode 1/n
Summary:
This is a stack to refactor GlobNode out of inode directory as much as possible.
I'm working on fine grained watchman subscription and i need to use GlobNode for globbing, which will have fs files depend on inode files. According to Katie, we want inode to be the top level.

Since the api im using won't require inode, the plan is to move the part that doesn't require inode into utils and have the inode glob search remain in inode

To split inode and tree implementation in this file, we need to make `evaluateImpl` and `evaluateRecursiveComponentImpl` stop referring to inodes, which is very hard without rewriting the methods because both methods do double recursion inside.

This code is hacky as it reuses runtime error for lines that we know will never reach during compile time. But I think this file is already using this hacky way multiple times https://www.internalfb.com/code/fbsource/[6cf48e74661705c6aefc18ebaecd7e229d71cb99]/fbcode/eden/fs/inodes/GlobNode.cpp?lines=122

I think with modern c++ we can probably get away this exception with constraints. But I'm not confident to do big refactor on this file for now

Reviewed By: kmancini

Differential Revision: D49900357

fbshipit-source-id: a769137238da78aba1da2720f9f1fc07f7e3ad6e