fix a FilteredBackingStore bug and document edge case
Summary:
We originally tried to pass in a FilteredObjectId into the underlying BackingStore's getBlob() function. This obviously won't work. This diff parses out the underlying ObjectID and passes that to the underlying BackingStore instead.

In addition, one might think that importManifestForRootId would need the same parsing logic for the Manifest parameter. However, any manifest we receive from Eden will not be "filtered", so we don't need to parse out the underlying Manifest and can simply pass that straight through. We do however have to parse the RootID we receive.

Reviewed By: genevievehelsel

Differential Revision: D49171303

fbshipit-source-id: 7c8e8de2c3fe1ed040bc171b5e79624dbd1a49b8