deubgexportstack: add "dataRef" for referring to large files
Summary:
We'd like to make `debugexportstack` "always" successful so ISL can always use
it and without dealing with error cases. To do that, we automatically turn
large files into a reference representation so it can be handled.

Note in case of a codemod, the exported "files" might be still too large to
consume. But it might still be okay to pass many files around and let's
optimize that case later.

Reviewed By: evangrayk

Differential Revision: D49842396

fbshipit-source-id: 302bef72d555bae2538b52c905eaab3870877271