Refactor unit tests
Summary:
Refactoring unit tests and a bit of the code to make the next diffs, which will include more unit tests for both modules, easier to review.

The goal of this is to provide everything needed for a test case inside the `GitExportTestData` struct. That's because the pattern of defining repos for scenarios that will be used in unit tests from both `partial_commit_graph` and `commit_rewrite` will continue in the next few diffs, so I wanted to avoid hardcoding and duplication.

Also made some small changes to the library code itself.

- Specifying the path types to `NonRootMPath` because there's not need to have it be generic to then cast it to `NonRootMPath` later.
- Using `anyhow::Result` instead of manually specifying error type.

Reviewed By: mitrandir77

Differential Revision: D49907742

fbshipit-source-id: 14b73685c293f766b136baa4eb39b704eeef48e0