Support heads per export path
Summary:
## Problem
If a export path was created by renaming another, the user might want to follow the history from the old path name. An example of this is `eden/mononoke` which was renamed in 2020 from `scm/mononoke` (D19722832).

See T164121717 and its attached diffs for more details about this problem.

## Solutions considered

**This diff implements solution #2.**

### 1. Try to automatically detect and follow renames
I spent some time implementing this, but there are a lot of edge cases and implementation would still take significant time and is likely to have issues with correctness (e.g. export what we don't want).

We decided it wasn't worth pursuing this given that this tool will be used by engineers.

### 2. Warn user and support passing rename revision manually (this diff)
During an export, we can (a) warn users when a changeset is likely creating the export path by renaming another one and (b) provide a way for the user to specify that we should export commits from a path but **only up to a specific changeset** (i.e. a HEAD commit).

This is what this diff and the one above implements.

## This diff
This diff implements support in the library, i.e. when export paths are given with specific head commits, it builds the proper `GitExportGraphInfo` and commits are copied properly given this `GitExportGraphInfo`.

The next diff adds arguments to the CLI so that this functionality can be used.

There's still one small improvement I want to do to the unit tests, but I figured I could publish this to speed up review while I update the tests (and possibly think of more test cases).

## TODO:
- Update unit tests to check copy_from field

Reviewed By: mitrandir77

Differential Revision: D49682305

fbshipit-source-id: 535f314c76214cfaa6aca90c9b48a4c7523b95de